requirements.txt
coverage==4.0
linecache2==1.0.0
nose==1.3.7
PyYAML==3.11
six==1.10.0
traceback2==1.4.0
unittest2==1.1.0
wheel==0.24.0
